requirements.txt
flask
torch
torchvision
opencv-python
mysql-connector-python
